Added compatability for certain client commands to send wanted strings
Common:
Added space for create_text_msg command to allow different command types
(this still needs work).

Server:
Removed the decode statement on the text_msg_handler as to avoid errors.

Client:
Made commands call text() to send the necessary strings to the server.
Put a placeholder ' ' in the text method call for command type sending
between calls (this needs some work still).

Currently, if the client successfully calls a command that sends a
message to the server (other than command type/id -- i.e. cd, get, put,
mget, mput) where the server receives the string and prints it to the
screen. In the future, it would be preferable to pick up the command
type and decode the message in a way retaining to that specific command
type.