Basic debugging catches
This is all that I have time to do right now. I pulled your code and ran
it and tested different commands. The only thing that I changed was in
the server code so that when you use the cd command, the directory
changes to whatever is sent instead of appending to the previous
directory -- this appending kept the changing of directories to a locked
forward path, meaning that the user could only go deeper into the
directories files and never back up. My change effectively removes this
restriction by removing the append and instead changing the directory to
what is sent -- meaning that if "Spencer Massey/Desktop" is sent, the
directory will not be changed due to the fact that this directory is not
valid.

Other catches:
~ When the user calls the normal command without changing any settings,
an error is received and the program crashes. I do not know the cause to
this but will look into it.
~ When the user toggles any setting and then calls the normal command,
the result is the toggling of the encryption and compression settings
but not the binary. Side note here: this should default the settings
back to what they were at the beginning instead of just toggling.

If I get any more extra time, I will take a loot at these issues and try
and move the project forward.