Fixed the stuff I saw last
The changes I made were to fix the problems I addressed in my last
commit. The program still runs completely on my machine, however, I do
not know if this extends to yours. The changes effect the program in
this way:

Normal now switches binary to true and the other two settings to false.
Normal will return false if none of the above settings are switched and
will return true if any of the settings are switched.

cd now allows for either full pathnames or paths that exist in the
current directory. For example:
(current directory is C:/Users/)
>cd /public
(current directory is C:/Users/public)
>cd C:/Users
(current directory is C:/Users)