Fixed my fixes and added password syntax check
I read the project guideline paper thingy and saw that it says that
normal doesn't effect binary, so I took that out. I also got the
re.match command working to check the syntax of the email you use for
your password. The password has to have one @ symbol followed by at
least one . to be accepted. Also, no whitespace is allowed.

IMPORTANT:
To run the program, you either need to add a new user and password that
is allowed (at least before in-class test-day) or comment out the 3
lines that check for the syntax of the password (probably gonna happen
for easier testing).